refactor:중복 코드 통합
